+++ MOSTComputers.Services.ProductRegister/Services/PromotionService.cs
@@ -41,27 +41,35 @@ internal sealed class PromotionService : IPromotionService
         return _promotionRepository.GetAllActive();
     }
 
-    public IEnumerable<Promotion> GetAllForProduct(uint productId)
+    public IEnumerable<Promotion> GetAllForProduct(int productId)
     {
+        if (productId <= 0) return Enumerable.Empty<Promotion>();
+
         return _promotionRepository.GetAllForProduct(productId);
     }
 
-    public IEnumerable<Promotion> GetAllForSelectionOfProducts(List<uint> productIds)
+    public IEnumerable<Promotion> GetAllForSelectionOfProducts(List<int> productIds)
     {
+        productIds = RemoveValuesSmallerThanOne(productIds);
+
         return _promotionRepository.GetAllForSelectionOfProducts(productIds);
     }
 
-    public IEnumerable<Promotion> GetAllActiveForSelectionOfProducts(List<uint> productIds)
+    public IEnumerable<Promotion> GetAllActiveForSelectionOfProducts(List<int> productIds)
     {
+        productIds = RemoveValuesSmallerThanOne(productIds);
+
         return _promotionRepository.GetAllActiveForSelectionOfProducts(productIds);
     }
 
-    public Promotion? GetActiveForProduct(uint productId)
+    public Promotion? GetActiveForProduct(int productId)
     {
+        if (productId <= 0) return null;
+
         return _promotionRepository.GetActiveForProduct(productId);
     }
 
-    public OneOf<uint, ValidationResult, UnexpectedFailureResult> Insert(ServicePromotionCreateRequest createRequest, IValidator<ServicePromotionCreateRequest>? validator = null)
+    public OneOf<int, ValidationResult, UnexpectedFailureResult> Insert(ServicePromotionCreateRequest createRequest, IValidator<ServicePromotionCreateRequest>? validator = null)
     {
         ValidationResult validationResult = ValidateTwoValidatorsDefault(createRequest, validator, _createRequestValidator);
 
@@ -71,9 +79,9 @@ internal sealed class PromotionService : IPromotionService
 
         createRequestInternal.PromotionAddedDate = DateTime.Today;
 
-        OneOf<uint, UnexpectedFailureResult> result = _promotionRepository.Insert(createRequestInternal);
+        OneOf<int, UnexpectedFailureResult> result = _promotionRepository.Insert(createRequestInternal);
 
-        return result.Match<OneOf<uint, ValidationResult, UnexpectedFailureResult>>(
+        return result.Match<OneOf<int, ValidationResult, UnexpectedFailureResult>>(
             id => id, unexpectedFailure => unexpectedFailure);
     }
 
@@ -92,13 +100,17 @@ internal sealed class PromotionService : IPromotionService
         return result.Match<OneOf<Success, ValidationResult, UnexpectedFailureResult>>(
             success => success, unexpectedFailure => unexpectedFailure);
     }
-    public bool Delete(uint id)
+    public bool Delete(int id)
     {
+        if (id <= 0) return false;
+
         return _promotionRepository.Delete(id);
     }
 
-    public bool DeleteAllByProductId(uint productId)
+    public bool DeleteAllByProductId(int productId)
     {
+        if (productId <= 0) return false;
+
         return _promotionRepository.DeleteAllByProductId(productId);
     }
 }
